Implemented Caching Functions --  makeCache Matrix and cacheSolve functions
The `makeCacheMatrix` function creates a special matrix object that can store its inverse. The `cacheSolve` function computes the inverse of this matrix if it hasn't already  been cached—otherwise, it retrieves the cached result to avoid redundant calculations. Thus, these functions work together to cache the inverse of a matrix, reducing computational cost.

Co-Authored-By: Nandita Dasgupta <nanditadg@aol.com>